re-organize code: radiation now in factory form, with diasbled option
