+++ cases/dycoms/dycoms.ini
@@ -75,7 +75,7 @@ sigma=0.00223
 beta=2.
 
 [time]
-endtime=43200.
+endtime=86400.
 dt=3.
 dtmax=60.
 savetime=1800
@@ -88,7 +88,7 @@ jday=160
 
 [stats]
 swstats=1
-sampletime=1800
+sampletime=900
 masklist=ql
 
 [radiation]
